Добавил библиотеку, стало намного интереснее играт
Если вы играли с одной цифрой, теперь их 5!!!!!
