+++ Lib/Const.cpp
@@ -4,3 +4,4 @@
 const int GAME_MODE = 1;
 const int REDACTOR_MODE = 100;
 const int KOLVO_KNOPOK = 5;
+
