+++ main.cpp
@@ -2,41 +2,21 @@
 #include "Lib\\TXLib.h"
 #include "Lib\\Const.cpp"
 #include "Lib\\lol.cpp"
+#include "Lib\\Menu.cpp"
 
 int mode = GAME_MODE;
-
-void Level_1 (HDC TV, HDC Chair, HDC Table, HDC Bed)
-{
-      txBitBlt (txDC(), 200, 200, 30, 30, TV);
-      txBitBlt (txDC(), 200, 400, 30, 30, TV);
-      txBitBlt (txDC(), 400, 200, 30, 30, TV);
-      txBitBlt (txDC(), 300, 200, 30, 30, Chair);
-      txBitBlt (txDC(), 200, 300, 30, 30, Chair);
-      txBitBlt (txDC(), 300, 300, 30, 30, Chair);
-      txBitBlt (txDC(), 600, 200, 30, 30, Table);
-}
 
 int main()
 {
 
     txCreateWindow(1000,720);
 
-    HDC tv = txLoadImage ("Icons\\tv.bmp");
-
-    HDC chair = txLoadImage ("Icons\\tv.bmp");
-
     HDC  TV = txLoadImage ("Icons\\tv.bmp");
     HDC  Chair = txLoadImage ("Icons\\tv.bmp");
     HDC  Table = txLoadImage ("Icons\\tv.bmp");
     HDC  Bed = txLoadImage ("Icons\\tv.bmp");
 
-    Knopka knopki_dlya_menu[KOLVO_KNOPOK];
-    knopki_dlya_menu[0] = {0, 0, "vanna", -1, tv};
-    knopki_dlya_menu[1] = {200, 0, "zal", -1, tv};
-    knopki_dlya_menu[2] = {400, 0, "kuxnua", -1, tv};
-    knopki_dlya_menu[3] = {600, 0, "spalnua", -1, tv};
-    knopki_dlya_menu[4] = {800, 0, "spalnua", -1, tv};
-
+    zap();
     while (!GetAsyncKeyState(VK_ESCAPE))
     {
         txBegin();
@@ -59,7 +39,7 @@ int main()
         {
             if (knopki_dlya_menu[nomer_knopki].risovatKartinku == 1)
             {
-                txBitBlt (txDC(), 500, 500, 30, 30, knopki_dlya_menu[nomer_knopki].kartinka, 0, 0);
+                txBitBlt (txDC(), knopki_dlya_menu[nomer_knopki].x+100, 500, 30, 30, knopki_dlya_menu[nomer_knopki].kartinka, 0, 0);
             }
         }
 
@@ -71,9 +51,6 @@ int main()
     {
         txDeleteDC (knopki_dlya_menu[nomer_knopki].kartinka);
     }
-
-    txDeleteDC (tv);
-    txDeleteDC (chair);
 
     txDeleteDC (TV);
     txDeleteDC (Chair);
