+++ readme.txt
@@ -1,0 +1,16 @@
+-------------------------------------------------------------------------------------
+                            Green toilet
+                                                readme
+
+1)    арта квартиры будет разбита на квадраты 
+(или пр€моугольники).
+ѕоэтому все иконки должны быть кратного 
+размера  (если атом - 16 * 16, то комод размером 
+80 * 32 - это норма, а 76 * 19 - нет). 
+
+2) артинок у каждого объекта две - та, что выводитс€ 
+в меню(может быть любого небольшого
+размера, 71 * 13 или 90 * 22 - не важно)
+и иконка, котора€ выводитс€ на карте
+
+-------------------------------------------------------------------------------------
